unit-test for ND implemented
