+++ tests_core/samples/test_data.csv
@@ -1,0 +1,4 @@
+x1,x2
+1.0,2.0
+3.0,4.0
+5.0,6.0
